PP-11393 Move Stripe model classes
We have models that represent objects in Stripe API responses in 2
different packages.

Move all the models to the `gateway/stripe/json` package so they are in
the same place, and make the naming consistent.
