+++ src/main/java/uk/gov/pay/connector/events/model/dispute/DisputeCreated.java
@@ -2,7 +2,7 @@ package uk.gov.pay.connector.events.model.dispute;
 
 import uk.gov.pay.connector.client.ledger.model.LedgerTransaction;
 import uk.gov.pay.connector.events.eventdetails.dispute.DisputeCreatedEventDetails;
-import uk.gov.pay.connector.gateway.stripe.response.StripeDisputeData;
+import uk.gov.pay.connector.gateway.stripe.json.StripeDisputeData;
 
 import java.time.Instant;
