+++ src/main/java/uk/gov/pay/connector/gateway/stripe/json/StripeCustomer.java
@@ -1,10 +1,10 @@
-package uk.gov.pay.connector.gateway.stripe.response;
+package uk.gov.pay.connector.gateway.stripe.json;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class StripeTokenResponse {
+public class StripeCustomer {
     @JsonProperty("id")
     private String id;
