+++ src/main/java/uk/gov/pay/connector/gateway/stripe/json/StripeToken.java
@@ -1,10 +1,10 @@
-package uk.gov.pay.connector.gateway.stripe.response;
+package uk.gov.pay.connector.gateway.stripe.json;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class StripeCustomerResponse {
+public class StripeToken {
     @JsonProperty("id")
     private String id;
